some refactoring && perlin on gpu

+++ src/terrain/utils/hash_random.cpp
@@ -19,11 +19,11 @@ namespace wgen {
 
         h = splitmix64(h);
 
-        const double u = static_cast<double>(h >> 11) * (1.0 / 9007199254740992.0);
-        const double angle = u * 2.0 * std::numbers::pi_v<double>;
+        const float u = static_cast<double>(h >> 11) * (1.0 / 9007199254740992.0);
+        const float angle = u * 2.0 * std::numbers::pi_v<double>;
         return {
-            static_cast<float>(std::cos(angle)),
-            static_cast<float>(std::sin(angle))
+            std::cos(angle),
+            std::sin(angle)
         };
     }
